requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.30.0
pandas>=2.0.0
sqlalchemy>=2.0.0
requests>=2.31.0

# Database drivers
pyodbc>=4.0.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
cx-Oracle>=8.3.0

# For demo database (SQLite is built-in to Python)
# No additional SQLite dependencies needed